revert: move back method `try_instantiate` in `DataFlowInstance`
