+++ zenoh-flow/src/runtime/dataflow/mod.rs
@@ -18,26 +18,18 @@ pub mod node;
 
 use std::collections::HashMap;
 use std::sync::Arc;
-use uhlc::HLC;
 use uuid::Uuid;
 
 use crate::model::record::{
-    DataFlowRecord, LinkRecord, OperatorRecord, SinkRecord, SourceRecord, ZFConnectorKind,
-    ZFConnectorRecord,
+    DataFlowRecord, LinkRecord, OperatorRecord, SinkRecord, SourceRecord, ZFConnectorRecord,
 };
 // use crate::model::dataflow::validator::DataflowValidator;
-use self::instance::runners::connector::{ZenohReceiver, ZenohSender};
-use self::instance::runners::Runner;
-use self::instance::{create_links, DataFlowInstance};
 use self::node::{OperatorFactory, SinkFactory, SourceFactory};
 use crate::model::descriptor::{InputDescriptor, OutputDescriptor};
-use crate::prelude::{Context, ErrorKind};
 use crate::runtime::RuntimeContext;
-use crate::traits::{self, Node};
+use crate::traits;
 use crate::types::NodeId;
-use crate::{zferror, Result as ZFResult};
-
-use super::InstanceContext;
+use crate::Result as ZFResult;
 
 /// TODO(J-Loudet) Documentation: a DataFlow is the structure that lets us generate, through the
 /// different factories, a DataFlowInstance. However, a DataFlow is not a "DataFlowFactory": it can
@@ -211,138 +203,6 @@ impl DataFlow {
             connectors,
             links,
             counter: 0,
-        })
-    }
-
-    /// TODO(J-Loudet) Improve documentation.
-    ///
-    /// Try instantiating the DataFlow2, hence generating a DataFlowInstance2.
-    pub async fn try_instantiate(self, hlc: Arc<HLC>) -> ZFResult<DataFlowInstance> {
-        let instance_context = Arc::new(InstanceContext {
-            flow_id: self.flow.clone(),
-            instance_id: self.uuid,
-            runtime: self.context.clone(),
-        });
-
-        let mut node_ids: Vec<NodeId> = Vec::with_capacity(
-            self.source_factories.len()
-                + self.operator_factories.len()
-                + self.sink_factories.len()
-                + self.connectors.len(),
-        );
-
-        node_ids.append(&mut self.source_factories.keys().cloned().collect::<Vec<_>>());
-        node_ids.append(&mut self.operator_factories.keys().cloned().collect::<Vec<_>>());
-        node_ids.append(&mut self.sink_factories.keys().cloned().collect::<Vec<_>>());
-        node_ids.append(&mut self.connectors.keys().cloned().collect::<Vec<_>>());
-
-        let mut links = create_links(&node_ids, &self.links, hlc.clone())?;
-
-        let ctx = Context::new(&instance_context);
-
-        let mut runners = HashMap::with_capacity(self.source_factories.len());
-        for (source_id, source_factory) in &self.source_factories {
-            let mut context = ctx.clone();
-            let (_, outputs) = links.remove(source_id).ok_or_else(|| {
-                zferror!(
-                    ErrorKind::IOError,
-                    "Links for Source < {} > were not created.",
-                    &source_id
-                )
-            })?;
-
-            let source = source_factory
-                .factory
-                .new_source(&mut context, &source_factory.configuration, outputs)
-                .await?;
-            let runner = Runner::new(source, context.inputs_callbacks, context.outputs_callbacks);
-            runners.insert(source_id.clone(), runner);
-        }
-
-        for (operator_id, operator_factory) in &self.operator_factories {
-            let mut context = ctx.clone();
-            let (inputs, outputs) = links.remove(operator_id).ok_or_else(|| {
-                zferror!(
-                    ErrorKind::IOError,
-                    "Links for Operator < {} > were not created.",
-                    &operator_id
-                )
-            })?;
-
-            let operator = operator_factory
-                .factory
-                .new_operator(
-                    &mut context,
-                    &operator_factory.configuration,
-                    inputs,
-                    outputs,
-                )
-                .await?;
-            let runner = Runner::new(
-                operator,
-                context.inputs_callbacks,
-                context.outputs_callbacks,
-            );
-            runners.insert(operator_id.clone(), runner);
-        }
-
-        for (sink_id, sink_factory) in &self.sink_factories {
-            let mut context = ctx.clone();
-            let (inputs, _) = links.remove(sink_id).ok_or_else(|| {
-                zferror!(
-                    ErrorKind::IOError,
-                    "Links for Sink < {} > were not created.",
-                    &sink_id
-                )
-            })?;
-
-            let sink = sink_factory
-                .factory
-                .new_sink(&mut context, &sink_factory.configuration, inputs)
-                .await?;
-            let runner = Runner::new(sink, context.inputs_callbacks, context.outputs_callbacks);
-            runners.insert(sink_id.clone(), runner);
-        }
-
-        for (connector_id, connector_record) in &self.connectors {
-            let session = instance_context.runtime.session.clone();
-            let node = match &connector_record.kind {
-                ZFConnectorKind::Sender => {
-                    let (inputs, _) = links.remove(connector_id).ok_or_else(|| {
-                        zferror!(
-                            ErrorKind::IOError,
-                            "Links for Sink < {} > were not created.",
-                            connector_id
-                        )
-                    })?;
-                    Some(
-                        Arc::new(ZenohSender::new(connector_record, session, inputs).await?)
-                            as Arc<dyn Node>,
-                    )
-                }
-                ZFConnectorKind::Receiver => {
-                    let (_, outputs) = links.remove(connector_id).ok_or_else(|| {
-                        zferror!(
-                            ErrorKind::IOError,
-                            "Links for Source < {} > were not created.",
-                            &connector_id
-                        )
-                    })?;
-                    Some(
-                        Arc::new(ZenohReceiver::new(connector_record, session, outputs).await?)
-                            as Arc<dyn Node>,
-                    )
-                }
-            };
-
-            let runner = Runner::new(node, vec![], vec![]);
-            runners.insert(connector_id.clone(), runner);
-        }
-
-        Ok(DataFlowInstance {
-            _instance_context: instance_context,
-            data_flow: self,
-            runners,
         })
     }
 }
